Changes to make my bridge compile
